Day 22 | Part 2 - SpaceCard Data Int -> Long & Investigated Solution Runtime

+++ implementation/src/main/kotlin/com/aoc/cards/SpaceCardDeckFactory.kt
@@ -9,7 +9,7 @@ class SpaceCardDeckFactory {
 
     companion object {
         fun default() = factoryOrder(10007)
-        fun factoryOrder(quantity: Int) = SpaceCardDeck((0 until quantity).map(::SpaceCard).toCollection(LinkedList()))
+        fun factoryOrder(quantity: Long) = SpaceCardDeck((0 until quantity).map(::SpaceCard).toCollection(LinkedList()))
     }
 
 }
